requirements.txt
Django==2.1.5
django-debug-toolbar==1.11
django-shortuuidfield==0.1.3
djangorestframework==3.9.3
djangorestframework-stubs==0.4.2
pytz==2019.1
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
